Bumped connector version and updated connector readme. Replaced "Float" with "Double" as a numeric type option.

+++ airbyte-integrations/bases/base-java-s3/src/main/java/io/airbyte/integrations/destination/s3/util/NumericType.java
@@ -8,7 +8,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 
 public enum NumericType {
 
-  FLOAT("Float"),
+  DOUBLE("Double"),
   DECIMAL("Decimal");
 
   private final String value;
